Major update

+++ src/actions/user.ts
@@ -1,10 +1,11 @@
 "use server";
 
+import { checkAuth, checkSelfAuthOrAdmin } from "@/lib/auth-utils";
 import { createCRUDMessage } from "@/lib/string.utils";
 import { CreateUserData, UpdateUserData } from "@/schemas/user";
 import { userService } from "@/services/user.service";
 import { ActionResultState } from "@/types";
-import { User } from "@prisma/client";
+import { Role, User } from "@prisma/client";
 import _ from "lodash";
 import { revalidatePath } from "next/cache";
 
@@ -14,6 +15,7 @@ export async function createUserAction(
   data: CreateUserData
 ): Promise<ActionResultState<User>> {
   try {
+    checkAuth(Role.ADMIN);
     const createDto = _.omit(data, ["confirmPassword"]);
     const user = await userService.create(createDto);
 
@@ -26,7 +28,9 @@ export async function createUserAction(
     };
   } catch (error: any) {
     console.log(error.message);
-    return { message: createCRUDMessage(dataName, "create", "failed") };
+    return {
+      message: createCRUDMessage(dataName, "create", "failed", error.message),
+    };
   }
 }
 
@@ -35,6 +39,7 @@ export async function updateUserAction(
   data: UpdateUserData
 ): Promise<ActionResultState<User>> {
   try {
+    checkSelfAuthOrAdmin(id);
     const user = await userService.update(id, {
       ...data,
     });
@@ -48,12 +53,33 @@ export async function updateUserAction(
     };
   } catch (error: any) {
     console.log(error.message);
-    return { message: createCRUDMessage(dataName, "update", "failed") };
+    return { message: createCRUDMessage(dataName, "update", "failed", error.message) };
+  }
+}
+
+export async function toggleUserActivationAction(
+  id: string,
+): Promise<ActionResultState<User>> {
+  try {
+    checkSelfAuthOrAdmin(id);
+    const user = await userService.toggleActivation(id);
+    revalidatePath("/users");
+    revalidatePath("/account");
+
+    return {
+      success: true,
+      data: user,
+      message: createCRUDMessage(dataName, "update", "success"),
+    };
+  } catch (error: any) {
+    console.log(error.message);
+    return { message: createCRUDMessage(dataName, "update", "failed", error.message) };
   }
 }
 
 export async function deleteUserAction(id: string): Promise<ActionResultState> {
   try {
+    checkSelfAuthOrAdmin(id);
     await userService.delete(id);
     revalidatePath("/users");
 
@@ -61,7 +87,7 @@ export async function deleteUserAction(id: string): Promise<ActionResultState> {
       success: true,
       message: createCRUDMessage(dataName, "delete", "success"),
     };
-  } catch (error) {
-    return { message: createCRUDMessage(dataName, "delete", "failed") };
+  } catch (error: any) {
+    return { message: createCRUDMessage(dataName, "delete", "failed", error.message) };
   }
 }
